configs: sdm660: add default mixer controls
When audioserver is killed and restarted, mixer controls
to set proper configuration are not issued if the stream being
opened after restart matches the default values in HAL.

This results in afe port start failure if the configuration
is unsupported by the end device.

Add default mixer controls for appropriate backends in
mixer paths of all SDM660 variants so that they are fired during
platform_init when audioserver is restarted.

CRs-Fixed: 2014040
Change-Id: I053faff32661f5e2db6f0e3b5ac190aeb3d83d24